# docs/openapi-blog-system-v1.0.yaml
openapi: 3.0.3
info:
  title: 博客系统 REST API
  description: |
    博客系统后端 REST API 规范，基于 PRD v1.0、架构设计 v1.0、数据库设计 v1.0。
    
    涵盖9个业务模块：Auth、User、Article、Comment、Category、Tag、Upload、Admin、Analytics，
    共49个API端点。
    
    ## 认证方式
    - Access Token: `Authorization: Bearer {access_token}`（有效期2小时）
    - Refresh Token: HttpOnly Cookie（有效期7天/30天）
    
    ## 统一响应格式
    ```json
    { "code": 0, "message": "success", "data": { ... } }
    ```
  version: 1.0.0
  contact:
    name: 架构师 · 高见远
    email: arch@example.com
servers:
  - url: https://api.blog.com/api/v1
    description: 生产环境
  - url: http://localhost:3000/api/v1
    description: 本地开发
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access Token（有效期2小时）
    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh Token（HttpOnly Cookie）
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: 业务错误码
          example: 40101
        message:
          type: string
          description: 错误消息
          example: 邮箱或密码错误
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: 字段级错误详情（可选）

    # === Auth ===
    RegisterRequest:
      type: object
      required: [email, password, username]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: zhou@example.com
        password:
          type: string
          minLength: 8
          description: 密码（≥8位，含字母+数字）
          example: Tech2026!
        username:
          type: string
          minLength: 3
          maxLength: 20
          pattern: '^[a-zA-Z0-9_]+$'
          description: 用户名（注册后不可改）
          example: zhoujinbu
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: zhou@example.com
        password:
          type: string
          example: Tech2026!
        remember_me:
          type: boolean
          default: false
          description: 延长Refresh Token至30天
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIs...
        refresh_token:
          type: string
          example: dGhpcyBpcyBhIHJlZn...
        expires_in:
          type: integer
          example: 7200
        user:
          $ref: '#/components/schemas/UserBasic'
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIs...
        expires_in:
          type: integer
          example: 7200

    # === User ===
    UserBasic:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: zhou@example.com
        username:
          type: string
          example: zhoujinbu
        status:
          type: string
          enum: [pending_verification, active, suspended, deleted]
          example: active
        avatar:
          type: string
          nullable: true
          example: https://cdn.blog.com/avatars/1.jpg
        blog_name:
          type: string
          nullable: true
          example: 周进步的技术博客
        bio:
          type: string
          nullable: true
          example: 后端工程师，专注分布式系统
        roles:
          type: array
          items:
            type: string
          example: [author]
    UserPublic:
      type: object
      properties:
        username:
          type: string
          example: zhoujinbu
        blog_name:
          type: string
          example: 周进步的技术博客
        bio:
          type: string
          example: 后端工程师，专注分布式系统
        avatar:
          type: string
          example: https://cdn.blog.com/avatars/1.jpg
        social_links:
          type: array
          items:
            $ref: '#/components/schemas/SocialLink'
        post_count:
          type: integer
          example: 12
        created_at:
          type: string
          format: date-time
    SocialLink:
      type: object
      properties:
        platform:
          type: string
          enum: [github, twitter, weibo, zhihu, linkedin, custom]
          example: github
        url:
          type: string
          example: https://github.com/zhoujinbu
        display_name:
          type: string
          nullable: true
          example: My Site
        sort_order:
          type: integer
          example: 0
    UpdateProfileRequest:
      type: object
      properties:
        blog_name:
          type: string
          minLength: 5
          maxLength: 50
          nullable: true
        bio:
          type: string
          maxLength: 200
          nullable: true
        avatar:
          type: string
          nullable: true
          description: 头像URL（通过Upload获取）
        social_links:
          type: array
          maxItems: 5
          items:
            type: object
            required: [platform, url]
            properties:
              platform:
                type: string
                enum: [github, twitter, weibo, zhihu, linkedin, custom]
              url:
                type: string
              display_name:
                type: string
                nullable: true

    # === Article ===
...
    ArticlePublicListItem:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        summary:
          type: string
          nullable: true
        cover_image:
          type: string
          nullable: true
        category:
          $ref: '#/components/schemas/CategoryBrief'
        tags:
          type: array
          items:
            type: string
        view_count:
          type: integer
        comment_count:
          type: integer
        like_count:
          type: integer
        word_count:
          type: integer
        published_at:
          type: string
          format: date-time
        author:
          $ref: '#/components/schemas/AuthorBrief'
...
    AuthorBrief:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        avatar:
          type: string
          nullable: true
        blog_name:
          type: string
          nullable: true

    # === Comment ===
...
    CategoryBrief:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 10
    UsernameParam:
      name: username
      in: path
      required: true
      schema:
        type: string
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: 注册
      description: 邮箱+密码注册，注册后发送验证邮件
      operationId: authRegister
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 注册成功，验证邮件已发送 }
                  data:
                    type: object
                    properties:
                      id: { type: integer, example: 1 }
                      email: { type: string, example: zhou@example.com }
                      username: { type: string, example: zhoujinbu }
                      status: { type: string, example: pending_verification }
        '409':
          description: 邮箱已被注册(40901) / 用户名已被占用(40902)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 参数校验错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 注册频率超限(42901)
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: 验证邮箱
      operationId: authVerifyEmail
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          description: 验证成功（同时返回JWT）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string }
                  data:
                    $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Token无效或已过期(40001)
        '404':
          description: 用户不存在(40401)
  /auth/login:
    post:
      tags: [Auth]
      summary: 登录
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string }
                  data:
                    $ref: '#/components/schemas/TokenResponse'
        '401':
          description: 邮箱或密码错误(40101) / 邮箱未验证(40102) / 账号冻结(40103)
        '429':
          description: 登录锁定(42902)
  /auth/refresh:
    post:
      tags: [Auth]
      summary: 刷新Token
      operationId: authRefresh
      security:
        - CookieAuth: []
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string }
                  data:
                    $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Refresh Token无效(40104) / 已撤销(40105)
  /auth/logout:
    post:
      tags: [Auth]
      summary: 登出
      operationId: authLogout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string }
                  data: { type: object, nullable: true }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: 忘记密码（发送重置邮件）
      operationId: authForgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: 无论邮箱是否存在都返回200（防枚举）
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: 重置密码
      operationId: authResetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: 重置成功
        '400':
          description: Token无效或已过期(40002)
        '422':
          description: 密码不符合强度(42201)
  /auth/change-password:
    post:
      tags: [Auth]
      summary: 修改密码（已登录用户）
      operationId: authChangePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [old_password, new_password]
              properties:
                old_password: { type: string }
                new_password: { type: string, minLength: 8 }
      responses:
        '200':
          description: 修改成功（撤销所有RefreshToken）
        '401':
          description: 当前密码错误(40106)
        '422':
          description: 新密码不符合强度(42201)

  # === User ===
  /users/me:
    get:
      tags: [User]
      summary: 获取当前用户信息
      operationId: getUserMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  data:
                    allOf:
                      - $ref: '#/components/schemas/UserBasic'
                      - type: object
                        properties:
                          social_links:
                            type: array
                            items:
                              $ref: '#/components/schemas/SocialLink'
                          post_count:
                            type: integer
                          created_at: { type: string, format: date-time }
                          updated_at: { type: string, format: date-time }
    patch:
      tags: [User]
      summary: 更新用户资料
      operationId: updateUserMe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: 更新成功
        '422':
          description: 校验错误(42204/42205/42206)
  /users/{username}:
    get:
      tags: [User]
      summary: 获取用户公开主页
      operationId: getUserPublic
      security: []
      parameters:
        - $ref: '#/components/parameters/UsernameParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  data:
                    $ref: '#/components/schemas/UserPublic'
        '404':
          description: 用户不存在(40402)
  /users/{username}/articles:
    get:
      tags: [User, Article]
      summary: 获取用户公开文章列表（读者端）
      operationId: getUserArticles
      security: []
      parameters:
        - $ref: '#/components/parameters/UsernameParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: category_slug
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [published_at, view_count], default: published_at }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ArticlePublicListItem'
                      total: { type: integer }
                      page: { type: integer }
                      per_page: { type: integer }
        '404':
          description: 用户不存在(40402)
...